refactor: activity detail layout

+++ app/src/main/java/com/example/wanderlusty/feature_explore_tourism/data/datasource/LocalDataSource.kt
@@ -232,7 +232,7 @@ object TourismDataSource : LocalDataSource {
                         getString("description"),
                         getString("image").split(",").map { it.trim() },
                         getString("rating").toDouble(),
-                        getString("review").toInt(),
+                        getString("review").toDouble(),
                         getString("type"),
                         getString("location"),
                         getTourOptionList(optJSONArray("tour_option"))
